requirements.txt
altair==4.2.2
attrs==22.2.0
backports.zoneinfo==0.2.1
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
filelock==3.10.0
fonttools==4.39.2
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
lit==15.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
protobuf==3.20.3
psutil==5.9.4
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
rich==13.3.2
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.17.0
six==1.16.0
smmap==5.0.0
streamlit-nightly==1.20.1.dev20230316
sympy==1.11.1
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ultralytics==8.0.54
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
zipp==3.15.0